requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
beautifulsoup4==4.12.3
    # via
    #   -r requirements.in
    #   html-sanitizer
cachetools==5.5.1
    # via google-auth
certifi==2024.12.14
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   pdfminer-six
    #   requests
colorama==0.4.6
    # via tqdm
contourpy==1.3.1
    # via matplotlib
cryptography==44.0.0
    # via pdfminer-six
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via pymongo
exceptiongroup==1.2.2
    # via anyio
filetype==1.2.0
    # via langchain-google-genai
fonttools==4.55.8
    # via matplotlib
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.24.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.160.0
    # via google-generativeai
google-auth==2.38.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.4
    # via
    #   -r requirements.in
    #   langchain-google-genai
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.1.1
    # via playwright
grpcio==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.70.0
    # via google-api-core
h11==0.14.0
    # via httpcore
html-sanitizer==2.5.0
    # via -r requirements.in
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.8.2
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kiwisolver==1.4.8
    # via matplotlib
langchain-core==0.3.33
    # via
    #   langchain-google-genai
    #   langchain-openai
    #   langgraph
    #   langgraph-checkpoint
langchain-google-genai==2.0.9
    # via -r requirements.in
langchain-openai==0.3.3
    # via -r requirements.in
langgraph==0.2.68
    # via -r requirements.in
langgraph-checkpoint==2.0.10
    # via langgraph
langgraph-sdk==0.1.51
    # via langgraph
langsmith==0.3.3
    # via langchain-core
lxml==5.3.0
    # via
    #   html-sanitizer
    #   lxml-html-clean
    #   python-docx
lxml-html-clean==0.4.1
    # via html-sanitizer
matplotlib==3.10.0
    # via -r requirements.in
mistune==0.8.4
    # via -r requirements.in
msgpack==1.1.0
    # via langgraph-checkpoint
numpy==2.2.2
    # via
    #   contourpy
    #   matplotlib
    #   pandas
openai==1.60.2
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.15
    # via
    #   langgraph-sdk
    #   langsmith
packaging==24.2
    # via
    #   langchain-core
    #   matplotlib
pandas==2.2.3
    # via -r requirements.in
pdfminer-six==20231228
    # via
    #   -r requirements.in
    #   pdfplumber
pdfplumber==0.11.4
    # via -r requirements.in
pillow==11.1.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pdfplumber
    #   pillow-heif
pillow-heif==0.21.0
    # via -r requirements.in
playwright==1.49.1
    # via -r requirements.in
proto-plus==1.26.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   google-generativeai
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   openai
pydantic-core==2.27.2
    # via pydantic
pyee==12.0.0
    # via playwright
pymongo==4.11
    # via -r requirements.in
pypandoc==1.15
    # via -r requirements.in
pypandoc-binary==1.15
    # via -r requirements.in
pyparsing==3.2.1
    # via
    #   httplib2
    #   matplotlib
pypdfium2==4.30.1
    # via pdfplumber
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9
    # via google-auth
schedule==1.2.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
slack-bolt==1.22.0
    # via -r requirements.in
slack-sdk==3.34.0
    # via
    #   -r requirements.in
    #   slack-bolt
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.6
    # via beautifulsoup4
tenacity==9.0.0
    # via langchain-core
tiktoken==0.8.0
    # via langchain-openai
tqdm==4.67.1
    # via
    #   google-generativeai
    #   openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   google-generativeai
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   python-docx
tzdata==2025.1
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via requests
zstandard==0.23.0
    # via langsmith
